Converge on one base Json type
