+++ packages/mds-repository/@types/index.ts
@@ -15,19 +15,6 @@
  */
 
 import { InsertResult, UpdateResult } from 'typeorm'
-import { Nullable } from '@mds-core/mds-types'
-
-// eslint-reason recursive declarations require interfaces
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface JsonArray extends Array<Json> {}
-
-export interface JsonObject {
-  [property: string]: Json
-}
-
-export type JsonValue = string | number | boolean | JsonArray | JsonObject
-
-export type Json = Nullable<JsonValue>
 
 export interface InsertReturning<T> extends InsertResult {
   raw: T[]
